% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjR2.R
\name{adjR2}
\alias{adjR2}
\title{adjusts R2}
\usage{
adjR2(PRS, trait, cov, nPRS = 1, prev = NA, binary = TRUE)
}
\arguments{
\item{PRS}{vector of PRS estimates}

\item{trait}{vector of trait. If binary, Cases=1, controls=0}

\item{cov}{covariate matrix}

\item{nPRS}{number of PRS in combination}

\item{prev}{trait prevalence}

\item{binary}{Binary or quantitative trait. Defaults to TRUE}
}
\description{
For binary traits, estimates R2 on the liability scale, pseudoR2 on the liability scale (and probit/logit) from Lee et al. 2012 "Better Coefficient".
For both quantitative and binary traits, if PRS is the result of a linear combination of multiple PRS, computes adjusted R2 accounting for greater model complexity
(Marquez-Luna et al. 2017)
}
