% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPRS.R
\name{adjustPRS}
\alias{adjustPRS}
\title{Scale/transform PRS}
\usage{
adjustPRS(
  IDs,
  PRS,
  anc.mat,
  rescale = FALSE,
  cluster = FALSE,
  pops = NA,
  k = 3
)
}
\arguments{
\item{IDs}{vector of IDS. Note: PRS and anc.mat IDs should match}

\item{PRS}{vector of PRS to be adjusted}

\item{anc.mat}{Matrix of PCAs or ancestry proportions for IDs}

\item{rescale}{Rescale within population?Defaults to FALSE}

\item{cluster}{Determine population clusters from anc.mat? Defaults to FALSE}

\item{pops}{Population labels for Ids. Defaults to NA. If rescaling, need to provide one of pops or cluster}

\item{k}{Number of clusters. Defaults to 3.}
}
\description{
Adjusts PRS by PCA or ancestry proportions (adjusts for N-1; i.e., if 4 PCs are provided, adjusts for 3). In the case of ancestry
proportions, this prevents variables from summing to 1.
Also scales PRS within populations. If populations are not definied, populations will be determined via agglomerative clustering.
Note that it is assumed that the IDs for the PRS and anc.mat are matched
}
