% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAC.R
\name{calcAC}
\alias{calcAC}
\title{calculate allele counts}
\usage{
calcAC(geno)
}
\arguments{
\item{geno}{data.frame, matrix, or vector of genotypes}
}
\value{
vector of allele counts
}
\description{
Calculates allele counts of genotypes in a data.frame, matrix, or vector
}
