% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAF.R
\name{calcAF}
\alias{calcAF}
\title{calculate allele frequencies}
\usage{
calcAF(geno)
}
\arguments{
\item{geno}{data.frame, matrix, or vector of genotypes}
}
\value{
vector of allele frequencies
}
\description{
Calculates allele frequencies of genotypes in a data.frame, matrix, or vector
}
