% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNobs.R
\name{calcNobs}
\alias{calcNobs}
\title{calculate number of chromosomal observations}
\usage{
calcNobs(geno)
}
\arguments{
\item{geno}{data.frame, matrix, or vector of genotypes}
}
\value{
vector of chromosomal observations
}
\description{
calculate number of chromosomal observations in a data.frame, matrix, or vector
}
