% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPRS.R
\name{calcPRS}
\alias{calcPRS}
\title{Calculate polygenic risk score}
\usage{
calcPRS(
  mgData,
  vcfData,
  imputed = FALSE,
  CGAT = FALSE,
  AF = FALSE,
  af.cutoff = NA,
  prune = FALSE,
  kb = 1000,
  thresh = FALSE,
  pval = 5e-08
)
}
\arguments{
\item{mgData}{GWAS summary statistics, can be data.frame or filepath}

\item{vcfData}{File path to VCF, BCF, BED}

\item{imputed}{Is data imputed? Defaults to FALSE}

\item{CGAT}{Keep CG/AT sites? Defaults to FALSE}

\item{AF}{filter by allele frequencies? Defaults to FALSE}

\item{af.cutoff}{Allele frequency difference cutoff. Defaults to NA}

\item{prune}{Prune by kb window? defaults to false}

\item{kb}{kb window for prune}

\item{thresh}{filter by pvalue threshold? Defaults to FALSE}

\item{pval}{pvalue threshold. Defaults to 5E-08}
}
\value{
returns data.frame with columns ID, PRS_RAW, PRS_AVG, PRS_SCALED. If save is TRUE, also outputs
text file with model SNPs and a VCF file of SNPs included in model.
}
\description{
Calculates polygenic risk score with snps present in mgData object (merged target genotype + GWAS sumstats)
Can be used to validate pre-trained PRS. Also implements a simple pruning/thresholding procedure using a windowing procedure.
}
\examples{
 \dontrun{
 calcPRS(sumstats, "data.vcf.gz", prefix="cohort", save=TRUE)
 }
}
