% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPRS2.R
\name{calcPRS2}
\alias{calcPRS2}
\title{Calculate PRS, version 2}
\usage{
calcPRS2(geno.file, sumstats.file, snp.col, a1.col, beta.col, plink = "plink")
}
\arguments{
\item{geno.file}{path to genotype file}

\item{sumstats.file}{path to sumstats file}

\item{snp.col}{name of snp column}

\item{a1.col}{name of a1 column}

\item{beta.col}{name of beta column}

\item{plink}{path to plink}
}
\description{
Estimates a PRS using PLINK --score function given paths to genotype file and sumstats file,
i.e. file of posterior effect sizes from PRS-CS
}
