% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAF.R
\name{checkAF}
\alias{checkAF}
\title{check for allele frequencies}
\usage{
checkAF(ref.file, bcftools = "bcftools")
}
\arguments{
\item{ref.file}{file path of genotype data}

\item{bcftools}{path to bcftools}
}
\description{
check if allele frequencies are already calculated
}
