% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkINPUT.R
\name{checkINPUT}
\alias{checkINPUT}
\title{Check input}
\usage{
checkINPUT(
  trait,
  PRS,
  IDs = NULL,
  folds = NULL,
  cov = NULL,
  anc.mat = NULL,
  pops = NULL
)
}
\arguments{
\item{trait}{vector of trait values}

\item{PRS}{vector of PRS values}

\item{IDs}{vector of subject IDs. Defaults to NULL}

\item{folds}{vector of fold membership for cross-validation. Defaults to NULL}

\item{cov}{covariate matrix. Can be matrix or data.frame. Defaults to NULL}

\item{anc.mat}{ancestry matrix. Can be PCA or ancestry proportions. Defaults to NULL}

\item{pops}{vector of population labels. Defaults to NULL}
}
\description{
Checks phenotype data input (trait, PRS, covariate matrix, ancestry matrix). Creates and returns a dataframe.
}
