% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpPRS.R
\name{clumpPRS}
\alias{clumpPRS}
\title{Clump+Thresh PRS}
\usage{
clumpPRS(
  target.file,
  ref.file,
  sumstats.file,
  snp.col,
  chr.col,
  bp.col,
  a1.col,
  beta.col,
  pval.col,
  plink = "plink",
  kb = 250,
  R2 = 0.1,
  pval = 5e-08
)
}
\arguments{
\item{target.file}{prefix of target plink fileset}

\item{ref.file}{prefix of LD ref plink fileset}

\item{sumstats.file}{path to summary stats file}

\item{snp.col}{name of SNP column}

\item{chr.col}{chromosome column}

\item{bp.col}{pos/basepair column}

\item{a1.col}{A1/effect allele/alt allele column}

\item{beta.col}{beta column}

\item{pval.col}{pvalue column}

\item{plink}{path to plink}

\item{kb}{size of LD window. Default is 250 kb}

\item{R2}{R2 parameter for LD. Default is 0.1}

\item{pval}{P-value parameter for thresholding. Default is 5E-08}
}
\description{
Uses PLINK to obtain a P+T/C+T set of summary sttatitics given an LD reference file, target genotype file, GWAS sumstats, R2, kb window, and pvalue
parameters. These paramters can be learned by writing a loop or using PRSice-2, or a combination of the two. The resulting pruned
file can then be used with calcPRS2() to estimate a PRS.
}
