% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePRS.R
\name{combinePRS}
\alias{combinePRS}
\title{Combine PRS}
\usage{
combinePRS(
  IDs,
  PRS1,
  w1,
  PRS2,
  w2,
  PRS3 = 0,
  w3 = 0,
  PRS4 = 0,
  w4 = 0,
  intercept = 0
)
}
\arguments{
\item{IDs}{vector of IDS matching PRS1,PRS2, etc.}

\item{PRS1}{PRS1 vector}

\item{w1}{weight/coefficient for PRS1}

\item{PRS2}{PRS2 vector}

\item{w2}{weight/coefficient for PRS2}

\item{PRS3}{PRS3 vector (optional)}

\item{w3}{weight/coefficient for PRS3 (optional)}

\item{PRS4}{PRS4 vector (optional)}

\item{w4}{weight/coefficient for PRS4 (optional)}

\item{intercept}{intercept for linear combination. Defaults to 0}
}
\description{
Combines PRS using provided weights/coefficients and intercept
}
