% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePRS.R
\name{comparePRS}
\alias{comparePRS}
\title{Compare PRS models}
\usage{
comparePRS(PRS1, PRS2, trait, folds, cov, binary = FALSE)
}
\arguments{
\item{PRS1}{first PRS to be tested}

\item{PRS2}{second PRS to be tested}

\item{trait}{vector of trait. If binary, cases=1, controls=0}

\item{folds}{vector of assigned folds for each subject.}

\item{cov}{covariate matrix}

\item{binary}{is the trait binary? defaults to FALSE}
}
\description{
Compares PRS models using the Williams test from the psych package as described by Pain et al. 2021. Will report Base vs PRS1,
Base vs. PRS2, PRS1 vs. PRS2
}
