% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfData-class.R
\name{concat.vcfData}
\alias{concat.vcfData}
\title{Concatenate vcfData objects}
\usage{
concat.vcfData(x, y)
}
\arguments{
\item{x}{vcfData object 1}

\item{y}{vcfData object 2}
}
\value{
vcfData object
}
\description{
concatenate vcfData objects. Checks if number of subjects is the same. If both
have allele frequencies, they will be combined via c()
}
