% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatData.R
\name{concatData}
\alias{concatData}
\title{Concatenate data split by chromosome}
\usage{
concatData(filename)
}
\arguments{
\item{filename}{base name of file}
}
\value{
data.frame with concatenated data
}
\description{
Concatenates data split by chromosome, i.e. GWAS data.
Provide fie name with a space for chromosome number.
}
\examples{
\dontrun{
concatData("GWAS.results.chr .txt")
}
}
