% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatGENO.R
\name{concatGeno}
\alias{concatGeno}
\title{Concatenate genotype files}
\usage{
concatGeno(
  geno.list,
  prefix = "output",
  file.type = c("vcf", "bed", "bcf"),
  plink = "plink",
  bcftools = "bcftools"
)
}
\arguments{
\item{geno.list}{Text file with genotype filepaths or vector with file names}

\item{prefix}{Prefix, defaults to output}

\item{file.type}{Select one of vcf, bed, bcf}

\item{plink}{Path to plink}

\item{bcftools}{Path to bcftools}
}
\description{
Wrapper function for concatenating genotype files
}
