% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converGeno.R
\name{convertGENO}
\alias{convertGENO}
\title{convert genotype files}
\usage{
convertGENO(
  geno.file,
  from = c("bed", "bcf", "vcf"),
  to = c("bed", "bcf", "vcf"),
  prefix = "output"
)
}
\arguments{
\item{geno.file}{filepath to genotype data}

\item{from}{Filetype of original genotype data}

\item{to}{Desired filetype}

\item{prefix}{Desired prefix. Defaults to output}
}
\description{
Wrapper function for handling file conversions
}
