% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empPVAL.R
\name{empPVAL}
\alias{empPVAL}
\title{empirical p-value for PRS}
\usage{
empPVAL(PRS, trait, cov, perms = 1000)
}
\arguments{
\item{PRS}{vector of PRS estimates}

\item{trait}{vector of trait}

\item{cov}{covariate matrix}

\item{perms}{Number of permuations. Defaults to 1000}
}
\description{
Generates empirical p-value via N permuations. Adds 1 to both numerator and denominator to preven emp.p = 0.
}
