% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalPRS.R
\name{evalPRS}
\alias{evalPRS}
\title{evaluate PRS}
\usage{
evalPRS(PRS, trait, folds, cov, binary = FALSE)
}
\arguments{
\item{PRS}{vector of PRS estimates}

\item{trait}{vector of trait. If binary, set cases=1, controls=0}

\item{folds}{vector of assigned folds for each subject.}

\item{cov}{covariate matrix}

\item{binary}{false}
}
\description{
Performs more detailed evaluation PRS than the testPRS() function. Will report mean pvals, R2/pseudoR2, MSE/RMSE OR
AUC/accuracy/sensitivty/specificity (binary traits). R2/pseudoR2 is partial R2 obtained by Full.R2 - Base.R2. Pval is p-value of PRS in full model
}
