% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRELS.R
\name{filterRELS}
\alias{filterRELS}
\title{Remove close relatives}
\usage{
filterRELS(
  kin.mat,
  pairwise = FALSE,
  degree = "3rd",
  plink.file,
  plink = "plink",
  FID = TRUE
)
}
\arguments{
\item{kin.mat}{kinship matrix. Can be matrix or pairwise relationships}

\item{pairwise}{Pairwise or matrix? Defaults to FALSE}

\item{degree}{1st,2nd, or 3rd degree threshold. Deafults to 3rd degree.}

\item{plink.file}{Path to plink fileset}

\item{plink}{path to plink binary}

\item{FID}{include FIDs in output? Defaults to TRUE}
}
\description{
Simple Algorithm for removing relatives. Provide kinship matrix and algorithm will quickly identify an optimal set of relatives to remove.
First, highly connected subjects are determined by counting the number of pairwise comparisons they appear in (after filtering)
Each highly connected subject is removed, in order of number of connections, until all pairwise comparisons contain unique subjects
The remaining pairwise comparisons are resolved via genotype missingness rate.
NOTE: input can be kinship matrix where rownames are subject ids OR three column data.frame of pairwise relationships wich columns
ID1, ID2, kinship coefficient
}
