% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSUBJECTS.R
\name{filterSUBJECTS}
\alias{filterSUBJECTS}
\title{Filter subjects}
\usage{
filterSUBJECTS(IDs, INPUT)
}
\arguments{
\item{IDs}{vector of ids}

\item{INPUT}{vector of values for filtering}
}
\description{
Filter subjects by providing input data for each subject (i.e. heterozygosity, PC, ancestry proportions, etc.). Subjects will be idenfited that
exceed 3*SD from the mean of the provided data. Note that filterRELS() identifies relatives
}
