% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_CGAT.R
\name{findCGAT}
\alias{findCGAT}
\title{find CG/AT sites}
\usage{
findCGAT(ref, alt)
}
\arguments{
\item{ref}{ref allele}

\item{alt}{alt allele}
}
\value{
TRUE or FALSE
}
\description{
Flags CG/AT sites as they are strand ambiguous
}
\examples{
findCGAT("C", "G")
}
