% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_allele.R
\name{flip_allele}
\alias{flip_allele}
\title{flip allele to opposite strand}
\usage{
flip_allele(x)
}
\arguments{
\item{x}{A,T,C, or G}
}
\value{
flipped allele
}
\description{
flip allele to opposite strand
}
\examples{
flip_allele("A")
}
