% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_dosage.R
\name{flip_dosage}
\alias{flip_dosage}
\title{Flip dosages}
\usage{
flip_dosage(dosage)
}
\arguments{
\item{dosage}{vector or matrix of genotype dosages}
}
\description{
"Flips" dosages when allele switch occurs
}
