% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gData-class.R
\name{gData}
\alias{gData}
\title{Class gData}
\usage{
gData(
  chr,
  pos,
  id,
  ref,
  alt,
  N,
  beta = NA,
  se = NA,
  pval = NA,
  af = NA,
  build = "hg19"
)
}
\arguments{
\item{chr}{chromosome vector}

\item{pos}{position vector in bp}

\item{id}{snp ids vector}

\item{ref}{ref/a2 allele vector}

\item{alt}{alt/a1 allele vector}

\item{N}{sample size vector}

\item{beta}{beta effect size, if data is from summary statistics. Defaults to NA}

\item{se}{standard error of the beta. Defaults to NA.}

\item{pval}{pvalue, if data is from summary statistics. Defaults to NA}

\item{af}{allele frequencies. Defaults to NA}

\item{build}{Genome build. Defaults to hg19}
}
\description{
Constructor function and related methods for S3 class gData
}
