% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AF.R
\name{get_AF}
\alias{get_AF}
\title{get allele frequencies}
\usage{
get_AF(geno.file, plink = "plink", bcftools = "bcftools")
}
\arguments{
\item{geno.file}{file path to genotype file}

\item{plink}{path to plink. Default assumes availavle in path}

\item{bcftools}{path to bcftools. Default assumes available in path}
}
\value{
data.frame with columns SNP and AF
}
\description{
calculates allele frequencies external to R using appropriate software.
For VCF/BCF files, will first check to see if INFO contains AF
}
