% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputed_dosage.R
\name{imputed_dosage}
\alias{imputed_dosage}
\title{get dosages from imputed data}
\usage{
imputed_dosage(geno)
}
\arguments{
\item{geno}{matrix or data frame of imputed genotypes}
}
\value{
matrix of imputed genotypes
}
\description{
If data is imputed, dosages will be included
}
