% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgData-class.R
\name{is.mgData}
\alias{is.mgData}
\title{Check of class mgData
Simple check if class mgData}
\usage{
is.mgData(x)
}
\arguments{
\item{x}{object}
}
\description{
Check of class mgData
Simple check if class mgData
}
