% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laiPRS.R
\name{LAiPRS}
\alias{LAiPRS}
\title{LAI partial PRS}
\usage{
LAiPRS(lai.vcf, anc.code, betas, snps = NA, norm = FALSE, PRS.ref = NA)
}
\arguments{
\item{lai.vcf}{filepath to annoated LAI vcf}

\item{anc.code}{code of ancestry from LAI}

\item{betas}{vector of betas for weights}

\item{snps}{vector of snp ids to keep in estimation. Defaults to NA (no filtering)}

\item{norm}{normalize PRS against references. Defaults to FALSE}

\item{PRS.ref}{vector of PRS for references. Defaults to NA}
}
\description{
Estimates partial PRS according to local ancestry using the annoated VCF generated by annoLAI().
Provide betas for SNPs in annotated VCF (ensure they match direction of effect, can use mgData() and matchBETA() functions).
Can normalize partial PRS using PRS of ancestry references (Marnetto et al. 2020). For this to be valid, this reference PRS needs to be done on the same
set of SNPs. When combining partial PRS, Marnetto et al. 2020 recomend weighting by ancestry proportions.
}
