% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBETA.R
\name{matchBETA}
\alias{matchBETA}
\title{Match betas with target}
\usage{
matchBETA(mgData, CGAT = FALSE, AF = FALSE, af.cutoff = NA)
}
\arguments{
\item{mgData}{mgData class object}

\item{CGAT}{Keep CG/AT sites? Defaults to FALSE}

\item{AF}{Run allele frequency check? Defaults to FALSE}

\item{af.cutoff}{allele frequency difference cutoff. Defaults to NA (which uses 2*SD)}
}
\description{
Match direction of effect with betas. Set betas of failed snps to 0 (i.e. not in reference, duplicates, mismatch).
}
