% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgData-class.R
\name{merge.gData}
\alias{merge.gData}
\alias{mgData}
\title{merge objects of class gData}
\usage{
\method{merge}{gData}(x, y, ..., by = c("CHROM", "POS"), all.x = TRUE, sort = FALSE)
}
\arguments{
\item{x}{gData object to merge}

\item{y}{gData object to merge}

\item{...}{further arguments passed to or from other methods.}

\item{by}{column for merging. Default is by chrom and pos}

\item{all.x}{Defaults to true}

\item{sort}{Defaults to FALSE}
}
\value{
object of class mgData
}
\description{
S3 method for merging class gData and contructor function for class mgData
}
