% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSUMSTATS.R
\name{parseSUMSTATS}
\alias{parseSUMSTATS}
\title{Parse GWAS sumstats}
\usage{
parseSUMSTATS(
  sumstats.file,
  snp.col,
  beta.col,
  pval.col,
  chrom.col,
  pos.col,
  ref.col,
  alt.col,
  af.col = NA,
  se.col = NA,
  N.col = NA,
  N = NA,
  OR = FALSE,
  build = "hg19",
  bim = NA
)
}
\arguments{
\item{sumstats.file}{Sumstats file}

\item{snp.col}{column name for snp IDs.}

\item{beta.col}{column name for betas.}

\item{pval.col}{column name for pvals.}

\item{chrom.col}{column name for chromosome number.}

\item{pos.col}{column name for physical position.}

\item{ref.col}{column name for reference/other allele.}

\item{alt.col}{column name for alt/effect allele.}

\item{af.col}{column name for effect allele frequency. Defaults to NA}

\item{se.col}{column name for standard error of the beta. Defaults to NA}

\item{N.col}{column name for sample size. defaults to NA}

\item{N}{sample size, if single value. Defaults to NA}

\item{OR}{effect sizes are odds ratios; convert to betas. Defaults to false}

\item{build}{Genome build. defaults to hg19}

\item{bim}{Bim file containing rsids, if needed. Defaults to NA}
}
\description{
Parse GWAS sumstats and return as gData object
}
