% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPRS.R
\name{plotPRS}
\alias{plotPRS}
\title{plot PRS distribution}
\usage{
plotPRS(trait, PRS, add.density = TRUE)
}
\arguments{
\item{trait}{trait vector. Can be binary or quantitative}

\item{PRS}{PRS vector}

\item{add.density}{If quantitative trait, color points by densiy? Defaults to TRUE}
}
\description{
Generates simple plots of PRS distribution. For quantitative trait, generates a scatter plot of TRAITxPRS.
For binary trait, generates a desnity plot of PRS stratified by status.
}
