% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPRS.R
\name{predictPRS}
\alias{predictPRS}
\title{Predict with PRS}
\usage{
predictPRS(IDs, PRS, trait, folds, cov, binary = FALSE)
}
\arguments{
\item{IDs}{vector of IDS for PRS, trait, folds, cov matrix, etc.}

\item{PRS}{vector of PRS estimates}

\item{trait}{vector of trait. If binary, cases=1, controls=0}

\item{folds}{vector of assigned folds for each subject.}

\item{cov}{covariate matrix}

\item{binary}{is the trait binary? defaults to FALSE}
}
\description{
Generates predictions using full (covariates+PRS) and base (covariates only) models across k folds where (N-N/k) is used to predict N/k subjects
Predictions can then be used for generating figures, if desired (i.e. ROC)
}
