% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsFOLDS.R
\name{prsFOLDS}
\alias{prsFOLDS}
\title{PRS folds}
\usage{
prsFOLDS(subjects, binary = FALSE, case_control = NA, folds = 10)
}
\arguments{
\item{subjects}{vector of subject IDS}

\item{binary}{Binary trait for stratified sampling. Defaults to FALSE}

\item{case_control}{Vector of case/control status. Defaults to NA}

\item{folds}{Number of folds. Defaults to 10}
}
\description{
Generates folds for PRS folds
}
\details{
For quantitative traits, assigns subjects to k folds randomly. For binary traits, assigns equal number of case/controls to k folds.
}
