% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read in VCF/BCF/BED}
\usage{
read_vcf(input.vcf, plink = "plink", bcftools = "bcftools")
}
\arguments{
\item{input.vcf}{filepath to genotype data}

\item{plink}{path to plink. Defaults to working directory}

\item{bcftools}{path to bcftools Defaults to working directory}
}
\value{
data.frame with vcf
}
\description{
Reads VCF/BCF/BED record into memory
}
\examples{
\dontrun{
read_vcf("data.vcf.gz")
}
}
