% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameSNPS.R
\name{renameSNPs}
\alias{renameSNPs}
\title{rename SNPs}
\usage{
renameSNPs(
  mgData,
  CGAT = FALSE,
  vcf.file = NA,
  prefix = "output",
  compress = TRUE,
  bgzip = "bgzip"
)
}
\arguments{
\item{mgData}{mgData object}

\item{CGAT}{keep CG/AT sites. defaults to FALSE}

\item{vcf.file}{file path to vcf file. Defaults to NA. If provided, will read in and write out vcf with updated ids.}

\item{prefix}{output prefix. Defaults to "output"}

\item{compress}{Compress VCF? defaults to true}

\item{bgzip}{Path to bgzip or gzip}
}
\description{
Takes advantage of the SNP matching routine in constructing mgData class object to write out SNP ids from target and gwas stats
for updating target SNP ids. Can also write out a filtered vcf file with new ids, if desired.
}
