% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPRS.R
\name{simPRS}
\alias{simPRS}
\title{Simulate PRS and a trait}
\usage{
simPRS(N = 500, binary = TRUE)
}
\arguments{
\item{N}{Number of subjects. Defaults to 500}

\item{binary}{Binary or quantitative trait. Defaults to TRUE}
}
\description{
Simple function for generating simulated traits for testing purposes
}
