% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_vcf.R
\name{subset_vcf}
\alias{subset_vcf}
\title{Subset vcf file at given position}
\usage{
subset_vcf(input.vcf, chr, pos, snp = NA, build = "hg19")
}
\arguments{
\item{input.vcf}{filepath to vcf file}

\item{chr}{chromosome interger}

\item{pos}{physical poisition}

\item{snp}{snp ID. Defaults to NA}

\item{build}{genome build. Defaults to hg19}
}
\value{
data.frame of vcf at given posiiton
}
\description{
Subsets vcf file via system call for bcftools. File needs to be indexed
and bcftools needs to be installed and in path
}
\examples{
\dontrun{
subset_vcf("data.vcf.gz", chr=2, pos=756111)
}
}
