% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAF.R
\name{testAF}
\alias{testAF}
\title{Test allele frequencies}
\usage{
testAF(af1, af2, cutoff = NA, align = NA, plot = FALSE)
}
\arguments{
\item{af1}{alelle frequencies from target cohort}

\item{af2}{allele frequencies from GWAS cohort}

\item{cutoff}{Desired abs(allele frequency difference) cutoff}

\item{align}{vector indicating SAME/SWITCH/FLIP/FLIP_SWITCH. Defaults to NA}

\item{plot}{Generates plot of allele frequencies. Defaults to FALSE}
}
\description{
Simple test of allele frequencies. Returns a vector indicating PASS/FAIL.
SNP is flagged as FAIL if abs(AF DIFF) > mean+2*SD. Only variants with mean difference < 0.5 are included in estimation of mean
If desired, user can provide a hard cutoff, i.e. an allele frequency difference of 0.2 (absolute value)
}
