% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPRS.R
\name{testPRS}
\alias{testPRS}
\title{test PRS}
\usage{
testPRS(PRS, trait, folds, cov, binary = FALSE)
}
\arguments{
\item{PRS}{vector of PRS estimates}

\item{trait}{vector of trait. If binary, cases=1, controls=0}

\item{folds}{vector of assigned folds for each subject.}

\item{cov}{covariate matrix}

\item{binary}{is the trait binary? defaults to FALSE}
}
\description{
Performs quick test of PRS by reporting mean Pearson's R between the trait and the prediction across K folds
}
