% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dosage.R
\name{to_dosage}
\alias{to_dosage}
\title{Convert genotypes to dosages}
\usage{
to_dosage(geno)
}
\arguments{
\item{geno}{Genotype data in a matrix, data.frame, or vector}
}
\value{
vector or matrix with genotype dosages
}
\description{
Converts genotypes, i.e. 1/0, to genotype dosages, i.e 1.
}
