% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfData-class.R
\name{vcfData}
\alias{vcfData}
\title{class vcfData}
\usage{
vcfData(vcf)
}
\arguments{
\item{vcf}{data.frame of full data when reading vcf into R}
}
\description{
Stores just the genotype data as a data.frame and calculates frequencies if AF=TRUE
}
