% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightPRS.R
\name{weightPRS}
\alias{weightPRS}
\title{Learn weights for linear combination of PRS}
\usage{
weightPRS(PRS1, PRS2, PRS3 = NA, PRS4 = NA, trait, folds, cov, binary = FALSE)
}
\arguments{
\item{PRS1}{vector for PRS1}

\item{PRS2}{vector for PRS2}

\item{PRS3}{vector for PRS3 (optional)}

\item{PRS4}{vector for PRS4 (optional)}

\item{trait}{vector of trait. If binary, cases=1, controls=0}

\item{folds}{vector of assigned folds for each subject.}

\item{cov}{covariate matrix}

\item{binary}{is the trait binary? defaults to FALSE}
}
\description{
Learn weights for linear combination of PRS. Two methods will be implemented: 1) fitting a regression model and 2)
learning the weights empirically where the weights sum to 1. The latter method appears to work better for admixed cohorts.
}
