% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Write VCF}
\usage{
write_vcf(vcf, out.file)
}
\arguments{
\item{vcf}{data.frame of vcf file in memory}

\item{out.file}{name of new vcf}
}
\value{
new vcf file
}
\description{
Wites VCF from vcf in memory
}
\examples{
\dontrun{
write_vcf2(vcf, "data.vcf")
}
}
