% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGWAS.R
\name{writeGWAS}
\alias{writeGWAS}
\title{export parsed GWAS sumstats}
\usage{
writeGWAS(sumstats, type = "PRS-CS", prefix = "output", by.chr = FALSE)
}
\arguments{
\item{sumstats}{gData object of GWAS data}

\item{type}{PRS-CS, PRSice2, or LDSR. Defaults to PRS-cS}

\item{prefix}{Desired prefix. Defaults to "output"}

\item{by.chr}{Write our by chromosome? Defaults to FALSE}
}
\description{
write out parsed GWAS sumstats in formats for PRSice2, PRS-CS, and LD score regression
}
