% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_file.R
\name{zip_file}
\alias{zip_file}
\title{compress file}
\usage{
zip_file(file.name, bgzip = TRUE)
}
\arguments{
\item{file.name}{Path to file for compression}

\item{bgzip}{Use bgzip.}
}
\value{
zipped file
}
\description{
compresses file using system call to gzip or bgzip.
bgzip needs to be installed and in path. Otherwise, use gzip
}
\examples{
\dontrun{
zip_file("data.vcf", bgzip=TRUE)
}
}
